base Dockerfile on os image, not intermediate upstream image
This is to remove a dependency on an image outside of our control. Doing so
forced the dependencies to be rebuilt, which meant a newer version of
Ansible was installed so some changes to the playbooks were necessary to
match the latest syntax.

The OS image has also been changed from ubuntu to debian-slim. This should
result in a smaller overall image, as it is intended to be smaller than the
stock Ubuntu image.